Creating collector package to make code more maintainable
